Extract shared structured logger
